requirements.txt
requests
tweepy
python-dotenv
playwright